fix: Add check for presence of Gemfile

+++ 20-2.7.4/init.sh
@@ -18,7 +18,9 @@ source /etc/profile.d/secret.sh
 
 # if any changes to Gemfile occur between runs (e.g. if you mounted the
 # host directory in the container), it will install changes before proceeding
-bundle check || bundle install --jobs 4
+if [ -f Gemfile ]; then
+  bundle check || bundle install --jobs 4
+fi
 
 if [ "$RAILS_ENV" == "production" ]; then
   bundle exec rake assets:precompile
