+++ 20-3.0.3/init.sh
@@ -10,7 +10,9 @@ fi
 
 # if any changes to Gemfile occur between runs (e.g. if you mounted the
 # host directory in the container), it will install changes before proceeding
-bundle check || bundle install --jobs 4
+if [ -f Gemfile ]; then
+  bundle check || bundle install --jobs 4
+fi
 
 if [ "$RAILS_ENV" == "production" ]; then
   bundle exec rake assets:precompile
